requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
flask==3.0.0
langchain==0.1.0
openai==1.7.2
python-dotenv==1.0.0
requests==2.31.0
webdriver-manager==4.0.1
streamlit==1.28.1
chromadb==0.4.18
sentence-transformers==2.2.2
beautifulsoup4
gunicorn





